+++ examples/yflow.php
@@ -26,23 +26,20 @@ $driver = match (random_int(1, 4)) {
 };
 printf("Use %s\n", $driver::class);
 
-function factorial(int $n): int
-{
-    return ($n <= 1) ? 1 : $n * factorial($n - 1);
-}
+$factorial = static function (int $n) use (&$factorial): int {
+    return ($n <= 1) ? 1 : $n * $factorial($n - 1);
+};
 
 /**
  * @return JobInterface<mixed, mixed>
  */
-function Ywrap(callable $func, callable $wrapperFunc): JobInterface
-{
+$Ywrap = static function (callable $func, callable $wrapperFunc): JobInterface {
     $wrappedFunc = static fn ($recurse) => $wrapperFunc(static fn (...$args) => $func($recurse)(...$args));
 
     return new YJob($wrappedFunc);
-}
+};
 
-function memoWrapperGenerator(callable $f): Closure
-{
+$memoWrapperGenerator = static function (callable $f): Closure {
     static $cache = [];
 
     return static function ($y) use ($f, &$cache) {
@@ -52,41 +49,38 @@ function memoWrapperGenerator(callable $f): Closure
 
         return $cache[$y];
     };
-}
+};
 
 /**
  * @return JobInterface<mixed, mixed>
  */
-function Ymemo(callable $f): JobInterface
-{
-    return Ywrap($f, 'memoWrapperGenerator');
-}
+$Ymemo = static function (callable $f) use ($Ywrap, $memoWrapperGenerator): JobInterface {
+    return $Ywrap($f, $memoWrapperGenerator);
+};
 
-function factorialGen(callable $func): Closure
-{
+$factorialGen = static function (callable $func): Closure {
     return static function (int $n) use ($func): int {
         return ($n <= 1) ? 1 : $n * $func($n - 1);
     };
-}
+};
 
-function factorialYMemo(int $n): int
-{
-    return Ymemo('factorialGen')($n);
-}
+$factorialYMemo = static function (int $n) use ($Ymemo, $factorialGen): int {
+    return $Ymemo($factorialGen)($n);
+};
 
-$factorialJob = static function (YFlowData $data): YFlowData {
-    printf("*... #%d - Job 1 : Calculating factorial(%d)\n", $data->id, $data->number);
+$factorialJob = static function (YFlowData $data) use ($factorial): YFlowData {
+    printf("*.... #%d - Job 1 : Calculating factorial(%d)\n", $data->id, $data->number);
 
     // raw factorial calculation
-    $result = factorial($data->number);
+    $result = $factorial($data->number);
 
-    printf("*... #%d - Job 1 : Result for factorial(%d) = %d\n", $data->id, $data->number, $result);
+    printf("*.... #%d - Job 1 : Result for factorial(%d) = %d\n", $data->id, $data->number, $result);
 
     return new YFlowData($data->id, $data->number);
 };
 
 $factorialYJobBefore = static function (YFlowData $data): YFlowData {
-    printf(".*.. #%d - Job 2 : Calculating factorialYJob(%d)\n", $data->id, $data->number);
+    printf(".*... #%d - Job 2 : Calculating factorialYJob(%d)\n", $data->id, $data->number);
 
     return new YFlowData($data->id, $data->number, $data->number);
 };
@@ -102,37 +96,44 @@ $factorialYJob = static function ($factorial) {
 };
 
 $factorialYJobAfter = static function (YFlowData $data): YFlowData {
-    printf(".*.. #%d - Job 2 : Result for factorialYJob(%d) = %d\n", $data->id, $data->number, $data->result);
+    printf(".*... #%d - Job 2 : Result for factorialYJob(%d) = %d\n", $data->id, $data->number, $data->result);
 
     return new YFlowData($data->id, $data->number);
 };
 
-$factorialYMemoJob = static function (YFlowData $data): YFlowData {
-    printf("..*. #%d - Job 3 : Calculating factorialYMemo(%d)\n", $data->id, $data->number);
+$factorialYMemoJob = static function (YFlowData $data) use ($driver, $factorialYMemo): YFlowData {
+    printf("..*.. #%d - Job 3 : Calculating factorialYMemo(%d)\n", $data->id, $data->number);
 
-    $result = factorialYMemo($data->number);
+    $driver->delay(3);
+    $result = $factorialYMemo($data->number);
 
-    printf("..*. #%d - Job 3 : Result for factorialYMemo(%d) = %d\n", $data->id, $data->number, $result);
+    printf("..*.. #%d - Job 3 : Result for factorialYMemo(%d) = %d\n", $data->id, $data->number, $result);
 
     return new YFlowData($data->id, $data->number);
 };
 
 $factorialYJobDeferBefore = static function (YFlowData $data) {
-    printf("...* #%d - Job 4 : Calculating factorialYJobDefer(%d)\n", $data->id, $data->number);
+    printf("...*. #%d - Job 4 : Calculating factorialYJobDefer(%d)\n", $data->id, $data->number);
 
     return new YFlowData($data->id, $data->number, $data->number);
 };
 
-$factorialYJobDefer = new YJob(static function ($factorial) {
-    return static function ($args) use ($factorial) {
+$factorialYJobDefer = new YJob(static function ($factorial) use ($driver) {
+    return static function ($args) use ($factorial, $driver) {
         [$data, $defer] = $args;
 
-        return $defer(static function ($complete, $async) use ($data, $defer, $factorial) {
+        return $defer(static function ($complete, $async) use ($data, $defer, $factorial, $driver) {
             if ($data->result <= 1) {
+                $delay = random_int(1, 3);
+                printf("...*. #%d - Job 4 : Step factorialYJobDefer(%d) with delay %d\n", $data->id, $data->number, $delay);
+                $driver->delay($delay);
                 $complete([new YFlowData($data->id, $data->number, 1), $defer]);
             } else {
-                $async($factorial([new YFlowData($data->id, $data->number, $data->result - 1), $defer]), static function ($result) use ($data, $complete) {
+                $async($factorial([new YFlowData($data->id, $data->number, $data->result - 1), $defer]), static function ($result) use ($data, $complete, $driver) {
                     [$resultData, $defer] = $result;
+                    $delay = random_int(1, 3);
+                    printf("...*. #%d - Job 4 : Step async factorialYJobDefer(%d) with delay %d\n", $data->id, $data->number, $delay);
+                    $driver->delay($delay);
                     $complete([new YFlowData($data->id, $data->number, $data->result * $resultData->result), $defer]);
                 });
             }
@@ -144,7 +145,50 @@ $factorialYJobDeferAfter = static function ($args) {
     [$data, $defer] = $args;
 
     return $defer(static function ($complete) use ($data, $defer) {
-        printf("...* #%d - Job 4 : Result for factorialYJobDefer(%d) = %d\n", $data->id, $data->number, $data->result);
+        printf("...*. #%d - Job 4 : Result for factorialYJobDefer(%d) = %d\n", $data->id, $data->number, $data->result);
+
+        $complete([new YFlowData($data->id, $data->number), $defer]);
+    });
+};
+
+$fibonacciYJobDeferBefore = static function (YFlowData $data) {
+    printf("....* #%d - Job 5 : Calculating fibonacciYJobDefer(%d)\n", $data->id, $data->number);
+
+    return new YFlowData($data->id, $data->number, $data->number);
+};
+
+$fibonacciYJobDefer = new YJob(static function ($fibonacci) use ($driver) {
+    return static function ($args) use ($fibonacci, $driver) {
+        [$data, $defer] = $args;
+
+        return $defer(static function ($complete, $async) use ($data, $defer, $fibonacci, $driver) {
+            if ($data->result <= 1) {
+                $delay = random_int(1, 3);
+                printf("....* #%d - Job 5 : Step fibonacciYJobDefer(%d) with delay %d\n", $data->id, $data->number, $delay);
+                $driver->delay($delay);
+                $complete([new YFlowData($data->id, $data->number, 1), $defer]);
+            } else {
+                $async($fibonacci([new YFlowData($data->id, $data->number, $data->result - 1), $defer]), static function ($result1) use ($data, $complete, $driver, $async, $fibonacci) {
+                    [$resultData1, $defer1] = $result1;
+                    $async($fibonacci([new YFlowData($data->id, $data->number, $data->result - 2), $defer1]), static function ($result2) use ($data, $complete, $driver, $resultData1) {
+                        [$resultData2, $defer2] = $result2;
+                        $delay = random_int(1, 3);
+                        printf("....* #%d - Job 5 : Step async fibonacciYJobDefer(%d) with delay %d\n", $data->id, $data->number, $delay);
+                        $driver->delay($delay);
+                        $fibResult = $resultData1->result + $resultData2->result;
+                        $complete([new YFlowData($data->id, $data->number, $fibResult), $defer2]);
+                    });
+                });
+            }
+        });
+    };
+});
+
+$fibonacciYJobDeferAfter = static function ($args) {
+    [$data, $defer] = $args;
+
+    return $defer(static function ($complete) use ($data, $defer) {
+        printf("....* #%d - Job 5 : Result for fibonacciYJobDefer(%d) = %d\n", $data->id, $data->number, $data->result);
 
         $complete([new YFlowData($data->id, $data->number), $defer]);
     });
@@ -158,7 +202,10 @@ $flow = Flow::do(static function () use (
     $factorialYMemoJob,
     $factorialYJobDeferBefore,
     $factorialYJobDefer,
-    $factorialYJobDeferAfter
+    $factorialYJobDeferAfter,
+    $fibonacciYJobDeferBefore,
+    $fibonacciYJobDefer,
+    $fibonacciYJobDeferAfter
 ) {
     yield [$factorialJob];
     yield [$factorialYJobBefore];
@@ -168,6 +215,9 @@ $flow = Flow::do(static function () use (
     yield [$factorialYJobDeferBefore];
     yield [$factorialYJobDefer, null, null, null, new DeferAsyncHandler()];
     yield [$factorialYJobDeferAfter, null, null, null, new DeferAsyncHandler()];
+    yield [$fibonacciYJobDeferBefore];
+    yield [$fibonacciYJobDefer, null, null, null, new DeferAsyncHandler()];
+    yield [$fibonacciYJobDeferAfter, null, null, null, new DeferAsyncHandler()];
 }, ['driver' => $driver]);
 
 for ($i = 1; $i <= 5; $i++) {