+++ src/AsyncHandler/DeferAsyncHandler.php
@@ -7,6 +7,8 @@ namespace Flow\AsyncHandler;
 use Flow\AsyncHandlerInterface;
 use Flow\Event;
 use Flow\Event\AsyncEvent;
+use Flow\Event\PoolEvent;
+use Flow\IpPool;
 
 /**
  * @template T
@@ -15,10 +17,21 @@ use Flow\Event\AsyncEvent;
  */
 final class DeferAsyncHandler implements AsyncHandlerInterface
 {
+    /**
+     * @var IpPool<T>
+     */
+    private IpPool $ipPool;
+
+    public function __construct()
+    {
+        $this->ipPool = new IpPool();
+    }
+
     public static function getSubscribedEvents()
     {
         return [
             Event::ASYNC => 'async',
+            Event::POOL => 'pool',
         ];
     }
 
@@ -26,11 +39,19 @@ final class DeferAsyncHandler implements AsyncHandlerInterface
     {
         $ip = $event->getIp();
         $job = $event->getJob();
+
+        $popIp = $this->ipPool->addIp($ip);
         $next = $job([$ip->data, $event->getDefer()]);
-        $next(static function ($result) use ($event) {
+        $next(static function ($result) use ($event, $popIp) {
             [$data] = $result;
             $callback = $event->getCallback();
             $callback($data);
+            $popIp();
         });
+    }
+
+    public function pool(PoolEvent $event): void
+    {
+        $event->addIps($this->ipPool->getIps());
     }
 }