+++ src/Driver/FiberDriver.php
@@ -30,6 +30,8 @@ use function count;
  */
 class FiberDriver implements DriverInterface
 {
+    use DriverTrait;
+
     /**
      * @var array<mixed>
      */
@@ -97,9 +99,9 @@ class FiberDriver implements DriverInterface
             };
         };
 
-        $defer = static function ($isTick) {
-            return static function (Closure|JobInterface $job) use ($isTick) {
-                return static function (Closure $next) use ($isTick, $job) {
+        $defer = static function ($isTick) use (&$fiberDatas) {
+            return static function (Closure|JobInterface $job) use ($isTick, &$fiberDatas) {
+                return static function (Closure $next) use ($isTick, $job, &$fiberDatas) {
                     $fiber = new Fiber(static function () use ($isTick, $job, $next) {
                         try {
                             $job(static function ($return) use ($isTick, $next) {
@@ -115,13 +117,22 @@ class FiberDriver implements DriverInterface
                     });
 
                     $fiber->start();
+
+                    $fiberDatas[] = [
+                        'fiber' => $fiber,
+                        'next' => static function ($return) {}, /*function ($return) use ($isTick, $next) {
+                            if ($isTick === false) {
+                                $next($return);
+                            }
+                        },*/
+                    ];
                 };
             };
         };
 
         $tick = 0;
         $fiberDatas = [];
-        while ($stream['ips'] > 0 or count($this->ticks) > 0) {
+        do {
             foreach ($this->ticks as [
                 'interval' => $interval,
                 'callback' => $callback,
@@ -132,30 +143,25 @@ class FiberDriver implements DriverInterface
                 }
             }
 
-            $nextIp = null;
-            do {
-                foreach ($stream['dispatchers'] as $index => $dispatcher) {
-                    $nextIp = $dispatcher->dispatch(new PullEvent(), Event::PULL)->getIp();
-                    if ($nextIp !== null) {
-                        $stream['dispatchers'][$index]->dispatch(new AsyncEvent(static function (Closure|JobInterface $job) use ($async) {
-                            return $async(false)($job);
-                        }, static function (Closure|JobInterface $job) use ($defer) {
-                            return $defer(false)($job);
-                        }, $stream['fnFlows'][$index]['job'], $nextIp, static function ($data) use (&$stream, $index, $nextIp) {
-                            if ($data instanceof RuntimeException and array_key_exists($index, $stream['fnFlows']) and $stream['fnFlows'][$index]['errorJob'] !== null) {
-                                $stream['fnFlows'][$index]['errorJob']($data);
-                            } elseif (array_key_exists($index + 1, $stream['fnFlows'])) {
-                                $ip = new Ip($data);
-                                $stream['ips']++;
-                                $stream['dispatchers'][$index + 1]->dispatch(new PushEvent($ip), Event::PUSH);
-                            }
+            foreach ($stream['dispatchers'] as $index => $dispatcher) {
+                $nextIps = $dispatcher->dispatch(new PullEvent(), Event::PULL)->getIps();
+                foreach ($nextIps as $nextIp) {
+                    $stream['dispatchers'][$index]->dispatch(new AsyncEvent(static function (Closure|JobInterface $job) use ($async) {
+                        return $async(false)($job);
+                    }, static function (Closure|JobInterface $job) use ($defer) {
+                        return $defer(false)($job);
+                    }, $stream['fnFlows'][$index]['job'], $nextIp, static function ($data) use (&$stream, $index, $nextIp) {
+                        if ($data instanceof RuntimeException && array_key_exists($index, $stream['fnFlows']) && $stream['fnFlows'][$index]['errorJob'] !== null) {
+                            $stream['fnFlows'][$index]['errorJob']($data);
+                        } elseif (array_key_exists($index + 1, $stream['fnFlows'])) {
+                            $ip = new Ip($data);
+                            $stream['dispatchers'][$index + 1]->dispatch(new PushEvent($ip), Event::PUSH);
+                        }
 
-                            $stream['dispatchers'][$index]->dispatch(new PopEvent($nextIp), Event::POP);
-                            $stream['ips']--;
-                        }), Event::ASYNC);
-                    }
+                        $stream['dispatchers'][$index]->dispatch(new PopEvent($nextIp), Event::POP);
+                    }), Event::ASYNC);
                 }
-            } while ($nextIp !== null);
+            }
 
             foreach ($fiberDatas as $i => $fiberData) { // @phpstan-ignore-line see https://github.com/phpstan/phpstan/issues/11468
                 if (!$fiberData['fiber']->isTerminated() and $fiberData['fiber']->isSuspended()) {
@@ -168,7 +174,7 @@ class FiberDriver implements DriverInterface
             }
 
             $tick++;
-        }
+        } while ($this->countIps($stream['dispatchers']) > 0 or count($this->ticks) > 0);
     }
 
     public function delay(float $seconds): void