+++ src/Event.php
@@ -27,7 +27,7 @@ final class Event
     /**
      * The PULL event occurs when Flow need a next IP to async process.
      *
-     * This event allows you to choose what IP come next from your pushed IPs and will be used for async process execution.
+     * This event allows you to choose what IPs come next from your pushed IPs and will be used for async process execution.
      *
      * @Event("Flow\Event\PullEvent")
      */
@@ -41,4 +41,13 @@ final class Event
      * @Event("Flow\Event\PopEvent")
      */
     public const POP = 'pop';
+
+    /**
+     * The POOL event occurs when Flow need to get the pool IPs to process.
+     *
+     * This event allows you to get the pool IPs to process.
+     *
+     * @Event("Flow\Event\PoolEvent")
+     */
+    public const POOL = 'pool';
 }