+++ src/IpStrategy/MaxIpStrategy.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Flow\IpStrategy;
 
 use Flow\Event;
+use Flow\Event\PoolEvent;
 use Flow\Event\PopEvent;
 use Flow\Event\PullEvent;
 use Flow\Event\PushEvent;
@@ -47,6 +48,7 @@ class MaxIpStrategy implements IpStrategyInterface
             Event::PUSH => 'push',
             Event::PULL => 'pull',
             Event::POP => 'pop',
+            Event::POOL => 'pool',
         ];
     }
 
@@ -63,13 +65,14 @@ class MaxIpStrategy implements IpStrategyInterface
      */
     public function pull(PullEvent $event): void
     {
-        if ($this->processing < $this->max) {
-            $ip = $this->dispatcher->dispatch($event, Event::PULL)->getIp();
-            if ($ip) {
+        $ips = $this->dispatcher->dispatch(new PullEvent(), Event::PULL)->getIps();
+        foreach ($ips as $ip) {
+            if ($this->processing < $this->max) {
                 $this->processing++;
+                $event->addIp($ip);
+            } else {
+                $this->dispatcher->dispatch(new PushEvent($ip), Event::PUSH);
             }
-
-            $event->setIp($ip);
         }
     }
 
@@ -80,5 +83,10 @@ class MaxIpStrategy implements IpStrategyInterface
     {
         $this->dispatcher->dispatch($event, Event::POP);
         $this->processing--;
+    }
+
+    public function pool(PoolEvent $event): void
+    {
+        $event->addIps($this->dispatcher->dispatch($event, Event::POOL)->getIps());
     }
 }